Rename action points permission file
